Move ramjet files to a package
