+++ ramjet/photometric_database/tests/resources/test_candlist_RADec.dat.txt
