+++ ramjet/photometric_database/tests/resources/test_data_directory/negative/negative1-R-1-0-100869.phot.cor.feather
