+++ ramjet/photometric_database/tests/resources/test_data_directory/tess/data_validations/tess2018206190142-s0001-s0001-0000000126890217-00106_dvr.xml
