+++ ramjet/photometric_database/tests/resources/test_data_directory/tess/lightcurves/tess2018206045859-s0001-0000000117544915-0120-s_lc.fits
