+++ ramjet/photometric_database/tests/resources/test_data_directory/tess/lightcurves/tess2018234235059-s0002-0000000150065151-0121-s_lc.fits
